remove misplaced android ops from sge core, fix initial test adapters
- Remove 31 android Ops files wrongly placed in sge/src/main/scalajvm/
  (they belong only in sge-jvm-platform-api, not in sge core)
- Fix 10 test files for generated API changes:
  SgeTestFixture, GestureDetectorTest, BezierTest, TableCellTest,
  VertexBufferObjectIss561Suite, TmjTileIdsCsvRedSuite,
  CumulativeDistributionTest, SaveDataCodecSuite,
  ObjLoaderFanRedSuite, ParticleEffectIoRedSuite

sge/compile: 0 errors. sge/Test/compile: 31 remaining (abstract method stubs).

+++ sge/src/test/scala/sge/graphics/g3d/particles/SaveDataCodecSuite.scala
@@ -205,7 +205,11 @@ class SaveDataCodecSuite extends munit.FunSuite {
         """},"indices":[]}],"unique":{}}"""
     val rd = ResourceData.fromJson[ParticleEffect](readFromString[Json](text))
     val sd = rd.saveData
-    assertEquals(sd.load[Int]("index").getOrElse(-1), 42, "java.lang.Integer value must load as an Int")
+    assertEquals(
+      sd.load[java.lang.Integer]("index").getOrElse(java.lang.Integer.valueOf(-1)),
+      java.lang.Integer.valueOf(42),
+      "java.lang.Integer value must load as an Int"
+    )
     assertEquals(sd.load[String]("name").getOrElse(""), "flame", "java.lang.String value must load as a String")
   }
